+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 IMPROVEMENTS:
 
 - Modify some parameters [GH-8]
+- Modify outputs.tf [GH-9]
 
 ## 1.1.0 (February 1, 2020)